prototype/OCR/requirements.txt
pandas
Pillow
PyPDF2
openpyxl
google-cloud-vision